Revert "Merge pull request #10 from east-01/KartControllerV3"
This reverts commit fd9dcbbf5e5a8ea836ec2a03d34f519e5a04f644, reversing
changes made to f6fd00bca4ed0d3ce1d0da0deec892b4d12d5e31.
