+++ Assets/Materials/Track.physicMaterial
@@ -7,8 +7,8 @@ PhysicMaterial:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_Name: Track
-  dynamicFriction: 0.6
-  staticFriction: 0.6
+  dynamicFriction: 0
+  staticFriction: 0
   bounciness: 0
   frictionCombine: 0
-  bounceCombine: 0
+  bounceCombine: 1
