+++ Assets/Prefabs/Kart/Bot.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e7f4da3d971f3514ea1449674dfa9614
+guid: 1e1f3ae22a344784197efe9bec91cd3c
 PrefabImporter:
   externalObjects: {}
   userData:
