--- Assets/Prefabs/Kart/Kart.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: f02595d1132c86b47bc3bb79439e612f
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
