--- Assets/Prefabs/Kart/Kart_Bot.prefab
@@ -1,174 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!1001 &7775119190005058117
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 3608140939005531148, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_Name
-      value: Kart_Bot (1)
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_RootOrder
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalPosition.y
-      value: -1.51
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalPosition.z
-      value: 17.81
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 7504855781678858529, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
---- !u!1 &6481934012597870153 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 3608140939005531148, guid: f02595d1132c86b47bc3bb79439e612f, type: 3}
-  m_PrefabInstance: {fileID: 7775119190005058117}
-  m_PrefabAsset: {fileID: 0}
---- !u!114 &8601565547112359052
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6481934012597870153}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8d0455521fcd0814785f8e259692b7d0, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  dotProductToTurn:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 1
-      inSlope: -0.0000020200557
-      outSlope: -0.0000020200557
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0
-      outWeight: 0.059012875
-    - serializedVersion: 3
-      time: 1
-      value: 0
-      inSlope: -2.1877894
-      outSlope: -2.1877894
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.042918444
-      outWeight: 0
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  dotProductToDriftTurn:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 1
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0
-      outWeight: 0
-    - serializedVersion: 3
-      time: 1
-      value: 0.004699707
-      inSlope: -1.9793236
-      outSlope: -1.9793236
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0.04506439
-      outWeight: 0
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  turnFactorToThrottle:
-    serializedVersion: 2
-    m_Curve:
-    - serializedVersion: 3
-      time: 0
-      value: 1
-      inSlope: -0.14114808
-      outSlope: -0.14114808
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0
-      outWeight: 0.09978541
-    - serializedVersion: 3
-      time: 1
-      value: 0.25
-      inSlope: 0
-      outSlope: 0
-      tangentMode: 0
-      weightedMode: 0
-      inWeight: 0
-      outWeight: 0
-    m_PreInfinity: 2
-    m_PostInfinity: 2
-    m_RotationOrder: 4
-  tfThresholdCalcFrequency: 0.33
-  turnFactorCount: 3
-  tfThresholdSingleBrake: 0.8
-  tfThresholdSingleDrift: 0.275
-  tfThresholdSingleThrottle: 0.15
-  turnInput: {x: 0, y: 0}
-  throttleInput: 0
-  driftInput: 0
-  boostInput: 0
-  checkpointIndex: 0
-  turnAmount: 0
-  turnFactor: 0
-  targetPosition: {x: 0, y: 0, z: 0}
-  directionToTarget: {x: 0, y: 0, z: 0}
-  dot: 0
-  cross: 0
-  turnLR: 0
-  turnValue: 0
-  throttleState: 
-  averageTrackSpeed: 0
-  stuck: 0
-  stuckLR: 0
-  tfThresholdMax: 0
-  tfThresholdBrake: 0
-  tfThresholdDrift: 0
-  tfThresholdThrottle: 0
