--- Assets/Prefabs/Kart/Kart_Bot.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 251c32129b1813f499c1910b81171fea
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
