--- Assets/Scripts/Camera/KartControllerFollow.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7d0c722f2e4a90f4c8bd3eda2349f128
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
