+++ Assets/Scripts/Camera/StrictFollow.cs
@@ -17,6 +17,10 @@ public class StrictFollow : MonoBehaviour
     {
 
 		float fov = this.fov;
+		KartController kc = subject.GetComponent<KartController>();
+		if(kc != null && kc.isActiveAndEnabled && kc.TrackSpeed > 0 && kc.IsBoosting() && kc.GetBoostAmount() > 0) {
+			fov *= 1.3f;
+		}
 
 		displayFov = Mathf.Lerp(displayFov, fov, fovInterpolationFactor*Time.deltaTime);
 		GetComponent<Camera>().fieldOfView = displayFov;
