+++ Assets/Scripts/Kart/KartController.cs
@@ -1,20 +1,17 @@
 using UnityEngine;
 using System;
-using System.Data.Common;
 // Kart Controller is NOT ALLOWED to use UnityEngine.InputSystem. See HumanDriver!
 
 /**
- * Kart ControllerV3 by Ethan Mullen
+ * Kart Controller by Ethan Mullen
  */
 public class KartController : MonoBehaviour
 {
 
-	/* ----- Settings variables ----- */
-	public Transform kartModel;
-
-	[Header("General")]
+	/* ### Settings variables ### */
+	[Header("Controls")]
 	public float inputDeadzone = 0.1f;
-	public float wheelSlideTorque = 60;
+	public float crashStateLength = 0.15f;
 
 	[Header("Speed")]
 	public float maxSpeed = 20f;
@@ -25,16 +22,21 @@ public class KartController : MonoBehaviour
 	public float kartTurnSpeed = 2f;
 	public AnimationCurve kartTurnPower;
 	public float steeringWheelTurnSpeed = 5f;
-	public Vector2 turnMultiplierRangeDriftMatch = new(1.5f, 2f); // Turn boost for drift that matches joystick
-	public Vector2 turnMultiplierRangeDriftDiffer = new(-0.2f, 0.35f); // Turn decrease for drift that doesn't match joystick
+
+	public Vector2 driftTurnDirectionMatchRange = new(1.5f, 2f);
+	public Vector2 driftTurnDirectionDifferRange = new(-0.2f, 0.35f);
 
 	[Header("Drift")]
-	public float driftAge = 4f; // The time (in seconds) it takes for a drift to reach its age
+	/** The time (in seconds) that it takes for a drift to reach its age.
+	 *  Drift performance worsens as it reaches its age. */
+	public float driftAge = 4f;
 	public float driftAngleMin = 0.225f;
 	public float driftAngleMax = 0.4f;
 
-	public float driftHopSpeedPercent = 0.1f; // The percent of max speed that the player is allowed to hop at
-	public float driftEngageSpeedPercent = 0.25f; // The percent of max speed that the player is allowed to engage a drift at
+	/** The percent of max speed that the player is allowed to hop at */
+	public float driftHopSpeedPercent = 0.1f;
+	/** The percent of max speed that the player is allowed to engage a drift at */
+	public float driftEngageSpeedPercent = 0.25f;
 	public float driftVerticalVelocity = 3f;
 
 	[Header("Boost")]
@@ -44,47 +46,66 @@ public class KartController : MonoBehaviour
 	public float passiveBoostDrain = 3f;
 	public float activeBoostDrain = 1.75f;
 
-	/* ----- Runtime variable ----- */
+	/* ### Runtime variable ### */
 	private KartStateManager stateMgr;
 	private Rigidbody rb;
+	private KartState state { get { return stateMgr.state; } }
 
 	/* Input variables */
-	[Header("Input"), SerializeField] private Vector2 turn;
-	[SerializeField] private float throttle;
-	[SerializeField] private bool boosting;
-	
-	[Header("Runtime fields")] public Vector3 up = new(0, 1, 0);
-	public Vector3 kartForward;
-	public int momentum;
-	public bool grounded; // Stores last update's grounded status
-	public float airtime;
-	public float timeSinceLastCollision;
-	public float steeringWheelDirection; // A [-1, 1] range float indicating the amount the steering wheel is turned and the direction.
-	public float boostAmount;
+	private Vector2 turn;
+	private float throttle;
 
-	public int driftDirection; // Indicates if we're in a left/right drift
+	/** A [-1, 1] range float indicating the amount the steering wheel is turned and the direction. */
+	private float steeringWheelDirection;
+
+	// Velocity/Speed
+	/** Kartforward keeps track of which direction the transform.forward vector should face.
+	    Initially, it will be set to Vector3.forward or (TODO) the direction that the
+		  spawnpoint dictates.
+		kartForward is used in velocity calculations. */
+	/* TODO: Traction, a fix to all collision problems.
+	   Make a variable called hasTraction that checks if the kart's velocity was
+	     the intended velocity. If it does, we have traction and hasTraction = true
+	   We'll use hasTraction for the velocity manipulation section. Two schools of
+	     thought right now:
+		 1. If hasTraction == false, don't change the velocity at all
+		 2. If hasTraction == false, attempt to guide the velocity in the direction that the 
+		      simulated wheels would. I.e., if the car were sliding backwards at a slight angle, 
+			  the car would turn to correct the car so that transform.forward = velocity.
+	   */
+	public Vector3 kartForward;
+	/** Momentum is updated along side kartForward, showing if we're rolling forward or backwards. */
+	[SerializeField] public int momentum;
+
+	public int driftDirection { get; private set; } // Indicates if we're in a left/right drift
 	private float driftTheta;
 	private float driftThetaTarget;
+	[Header("Runtime fields")]
 	public bool driftParticles; // True if driftparticles should be showing
 	private bool driftHopRequest;
+
+	private float boostAmount;
+	private bool boosting;
+	private bool lastActivelyBoosting;
+
+	private bool grounded; // Stores last update's grounded status
+	private float airtime;
+	private float timeSinceLastCollision;
 
     private void Start()
     {
 		rb = GetComponent<Rigidbody>();
 		stateMgr = GetComponent<KartStateManager>();
 
-		if(kartModel == null) Debug.LogWarning("KartController on \"" + gameObject.name + "\" doesn't have a kartModel assigned."); 
-
-		timeSinceLastCollision = 1000f;
-
 		// TODO: Make kart spawn at a spawn position and face a certain direction
 		kartForward = new Vector3(1, 0, 0); 
 		transform.forward = kartForward;
-
     }
 
     private void Update()
     {
+		DrawRays();
+
 		bool grounded = Grounded();
 		if(grounded) { 
 			airtime = 0;
@@ -93,7 +114,7 @@ public class KartController : MonoBehaviour
 			airtime += Time.deltaTime;	
 		}
 		this.grounded = grounded;
-		this.driftParticles = airtime <= 0.05f && stateMgr.state == KartState.DRIFTING && driftDirection != 0;
+		this.driftParticles = airtime <= 0.05f && state == KartState.DRIFTING && driftDirection != 0;
 
 		// Steering wheel direction modification
 		if(Mathf.Abs(turn.x) > 0) { 
@@ -107,11 +128,11 @@ public class KartController : MonoBehaviour
 		
 		// If we're drifting and airborne, change drift direction to match joystick
 		// Use airtime to ensure we maintain drift direction during small falls.
-		if(stateMgr.state == KartState.DRIFTING && airtime > 0.05f && Math.Abs(turn.x) >= inputDeadzone) 
+		if(state == KartState.DRIFTING && airtime > 0.05f && Math.Abs(turn.x) >= inputDeadzone) 
 			driftDirection = (int)Mathf.Sign(turn.x);
 
 		driftThetaTarget = 0;
-		if(stateMgr.state == KartState.DRIFTING && driftDirection != 0 && Mathf.Abs(turn.x) >= inputDeadzone) { 
+		if(state == KartState.DRIFTING && driftDirection != 0 && Mathf.Abs(turn.x) >= inputDeadzone) { 
 			driftThetaTarget = Mathf.Sign(driftDirection)*driftAngleMin;
 			if(SteeringWheelMatchesDrift) 
 				driftThetaTarget += steeringWheelDirection*(driftAngleMax-driftAngleMin);
@@ -132,91 +153,95 @@ public class KartController : MonoBehaviour
 		// Variable kartForward can never be 0, we get stuck.
 		if(kartForward.magnitude == 0) kartForward = transform.forward;
 
-		// Animate kartModel forward
-		if(kartModel != null)
-			kartModel.forward = RotateVectorAroundAxis(transform.forward, up, driftTheta);
-
 	}
 
-	private float theta;
 	private void FixedUpdate() 
 	{
 
-		Vector3 velNoY = rb.velocity;
-		velNoY.y = 0;
-
-		timeSinceLastCollision += Time.fixedDeltaTime;
-
+		/* Update kartForward and momentum- make transform follow kartForward */
 		// Momentum gets updated first, then kartForward second
-		momentum = TrackSpeed > 0.1f ? (Vector3.Dot(rb.velocity, transform.forward) >= 0 ? 1 : -1) : 0;
-
-		// Update theta
-		theta = kartTurnSpeed * steeringWheelDirection * kartTurnPower.Evaluate(SpeedRatio) * DriftTurnMultiplier * Time.fixedDeltaTime;
-		if(!grounded) theta /= 3;
-		if(momentum < 0) theta = -theta;
-
-		switch(stateMgr.state) {
-			case KartState.DRIVING:
-			case KartState.DRIFTING:
-
-				if(velNoY.magnitude > 0.5f) {
-					kartForward = momentum*velNoY.normalized;
-				} else {
-					Vector3 forwardNoY = transform.forward;
-					forwardNoY.y = 0;
-					if(forwardNoY.magnitude > 0.05f) kartForward = forwardNoY;
-				}
-
-				Vector3 throttleForce = (ActivelyBoosting ? 1f : throttle) * acceleration * transform.forward;
-				if(Mathf.Abs(throttle) > inputDeadzone && TrackSpeed <= CurrentMaxSpeed)
-					rb.AddForce(throttleForce, ForceMode.Acceleration);	
-
-				Vector3 turnForce = theta*1000f*-Vector3.Cross(rb.velocity.normalized, up);
-				rb.AddForce(turnForce, ForceMode.Acceleration);
-				
-				transform.forward = kartForward;
-
-				if(driftHopRequest) {
-					rb.AddForce(up * driftVerticalVelocity, ForceMode.VelocityChange);
-					driftHopRequest = false;
-				}
-				break;
-
-			case KartState.COLLISION:
-
-				float dot = Vector3.Dot(rb.velocity.normalized, transform.forward);
-				kartForward = transform.forward;
-				if((TrackSpeed < 0.1f*maxSpeed || Mathf.Abs(dot) > 0.975f) && timeSinceLastCollision > 0.05f) {
-					stateMgr.state = KartState.DRIVING;
-					momentum = (int)Mathf.Sign(dot);
-				} else { 
-					// Wheel sliding force
-					float minSR = 0f;
-					float sr = SpeedRatio;
-					if(sr < minSR) sr = minSR;
-					Vector3 rawTorque = wheelSlideTorque*Math.Max(Mathf.Abs(dot), sr)*-Vector3.Cross(rb.velocity.normalized, transform.forward);
-					Vector3 torque = up.normalized*Vector3.Dot(rawTorque, up);
-					rb.AddTorque(torque, ForceMode.Acceleration);
-				}
-
-				break;
+		momentum = Vector3.Dot(rb.velocity, transform.forward) >= 0 ? 1 : -1;
+		Vector3 kartForwardCandidate = rb.velocity;
+		kartForwardCandidate.y = 0;
+		if(CanChangeForward && kartForwardCandidate.magnitude > 0) { 
+			kartForward = rb.velocity.normalized*momentum;	
 		}
+		if(timeSinceLastCollision < crashStateLength) { 
+			kartForward = transform.forward;
+		}
+		
+		Vector3 transformForward = kartForward; // Default kartForward for state DRIVING
+		transformForward.y = 0;
+		transformForward = RotateVectorAroundAxis(transformForward, Vector3.up, driftTheta);
+		if(CanChangeForward && transformForward.magnitude > 0 && timeSinceLastCollision > crashStateLength) 
+			transform.forward = transformForward.normalized; // Use TrackSpeed > 1.15f to stop randomly turning at slow speeds
+
+		/** Crash state */
+		if(timeSinceLastCollision + Time.fixedDeltaTime >= crashStateLength && timeSinceLastCollision < crashStateLength) { 
+			// Reached end of crash state, make kart forward match velocity
+			kartForward = transform.forward;
+			Vector3 targetVelocity = kartForward*momentum*rb.velocity.magnitude;
+			rb.AddForce(targetVelocity-rb.velocity, ForceMode.VelocityChange);
+		}
+		this.timeSinceLastCollision += Time.fixedDeltaTime;
+
+		HandleVelocity();
+
+		lastActivelyBoosting = ActivelyBoosting;
 
 	}
 
-	public void OnCollisionEnter(Collision collision)
-	{
-		if(collision.collider.isTrigger) return;
+	//https://www.reddit.com/r/Unity3D/comments/psukm1/know_the_difference_between_forcemodes_a_little/
+	/** Takes the state of input and state manager and adds forces to the Rigidbody */
+	public void HandleVelocity() 
+	{ 
+		driftThetaTarget = 0;
 
-		for(int i = 0; i < collision.contactCount; i++) { 
-			Vector3 norm = collision.GetContact(i).normal;
-			if(Vector3.Dot(up, norm) < 0.5) { 
-				stateMgr.state = KartState.COLLISION;
-				rb.AddForce(-(rb.velocity.normalized*(Vector3.Dot(norm, rb.velocity)))*0.5f, ForceMode.VelocityChange);
-				return;
+		float throttle = this.throttle;
+		if(ActivelyBoosting) throttle = 1f;
+
+		/* Forward force application */
+		Vector3 throttleForce = kartForward * throttle * acceleration;
+		// Represents the velocity after throttleForce has been added using AddForce(ForceMode.Acceleration)
+		// Only used for VelocityChange calculations
+		Vector3 addedVelocity = kartForward + throttleForce*Time.fixedDeltaTime;
+
+		if(ActivelyBoosting && !lastActivelyBoosting) {
+			throttleForce *= 10;
+			// Make velocity == maxVelocity
+			rb.AddForce((addedVelocity.normalized*CurrentMaxSpeed)-rb.velocity, ForceMode.VelocityChange);
+		}
+
+		if(Mathf.Abs(throttle) > inputDeadzone) {
+			if(TrackSpeed <= CurrentMaxSpeed) { 
+				rb.AddForce(throttleForce, ForceMode.Acceleration);	
+			} else { 
+				rb.AddForce((addedVelocity.normalized*momentum*CurrentMaxSpeed)-rb.velocity, ForceMode.VelocityChange);
 			}
 		}
 
+		/* Turning */
+		float theta = kartTurnSpeed * steeringWheelDirection * Time.fixedDeltaTime;
+		theta *= kartTurnPower.Evaluate(SpeedRatio);
+		if(state == KartState.DRIFTING) {
+			float driftAgeRatio = stateMgr.timeInState/driftAge;
+			theta *= SteeringWheelMatchesDrift ? 
+				Mathf.Lerp(driftTurnDirectionMatchRange.x, driftTurnDirectionMatchRange.y, 1-driftAgeRatio) :
+				Mathf.Lerp(driftTurnDirectionDifferRange.x, driftTurnDirectionDifferRange.y, 1-driftAgeRatio);
+		}
+		if(!grounded) theta /= 3;
+		if(momentum < 0) theta = -theta;
+
+		if(theta != 0) {
+			/* Turn method #2: Rotate the velocity vector around the axis by theta, then add the difference
+					to the velocity. This should rotate the vector without changing velocity magnitude. */
+			rb.AddForce(RotateVectorAroundAxis(rb.velocity, Vector3.up, theta)-rb.velocity, ForceMode.VelocityChange);
+		}
+		
+		if(driftHopRequest) {
+			rb.AddForce(Vector3.up * driftVerticalVelocity, ForceMode.VelocityChange);
+			driftHopRequest = false;
+		}
 	}
 
 	/** The callback from KartStateManager indicating when we've changed state.
@@ -229,21 +254,48 @@ public class KartController : MonoBehaviour
 			case KartState.DRIVING:
 				break;
 			case KartState.DRIFTING:
-				if(!CanDriftEngage) {
+								
+				if(!CanDriftHop) {
 					stateMgr.state = KartState.DRIVING;
 					break;
 				}
 
 				driftHopRequest = true;
+				if(!CanEngageDrift) stateMgr.state = KartState.DRIVING;
+
 				break;
-			case KartState.COLLISION:
-				timeSinceLastCollision = 0;
-				boostAmount = 0;
+			case KartState.REVERSING:
 				break;
 			default:
 				break;
 		}
 	}
+
+	public void OnCollisionEnter(Collision collision)
+	{
+		bool wasGroundCollision = true;
+		for(int i = 0; i < collision.contactCount; i++) { 
+			if(Vector3.Dot(Vector3.up, collision.GetContact(i).normal) < 0.75) { 
+				wasGroundCollision = false;
+				break;
+			}
+		}
+
+		if(!wasGroundCollision) timeSinceLastCollision = 0f;
+
+	}
+
+	private Vector3 RemoveComponent(Vector3 vector, Vector3 normal)
+    {
+        // Calculate the projection of vector onto normal
+        float projection = Vector3.Dot(vector, normal);
+        Vector3 projectionVector = projection * normal;
+
+        // Subtract the projection from the original vector
+        Vector3 result = vector - projectionVector;
+
+        return result;
+    }
 
 	public Vector3 RotateVectorAroundAxis(Vector3 inputVector, Vector3 rotationAxis, float angleRadians)
     {
@@ -254,82 +306,71 @@ public class KartController : MonoBehaviour
 
 	public bool Grounded() 
 	{ 
-		float distance = 0.1f;
+		float distance = 0.05f;
 		Vector3 raycastOrigin = new(transform.position.x, GetComponent<BoxCollider>().bounds.min.y+0.001f, transform.position.z);
 		return Physics.Raycast(raycastOrigin, Vector3.down, distance);
 	}
 
 	/* Input */
-	public Vector2 TurnInput {
-		get { return turn; }
-		set {
-			turn = value;
-			if(turn.magnitude <= inputDeadzone) this.turn = Vector2.zero;
-		}
+	public Vector2 GetTurnInput() { return turn; }
+	public void SetTurnInput(Vector2 turn) {
+		this.turn = turn;
+		if(turn.magnitude <= inputDeadzone) this.turn = Vector2.zero;
 	}
 
-	public float ThrottleInput {
-		get { return throttle; }
-		set {
-			this.throttle = value;
-		}
+	public float GetThrottleInput() { return throttle; }
+	public void SetThrottleInput(float throttle) {
+		this.throttle = throttle;
 	}
 
-	public bool DriftInput { 
-		get { return stateMgr.state == KartState.DRIFTING; } 
-		set {
-			if(value && stateMgr.state == KartState.DRIVING && Grounded() && !ActivelyBoosting)
+	public void SetReverseInput(float reverse) {
+		this.throttle = -reverse;
+	}
+
+	public bool IsDrifting() { return state == KartState.DRIFTING; }
+	public void SetDriftInput(bool buttonDown) {
+		if(buttonDown) { 
+			if(state == KartState.DRIVING && Grounded() && !ActivelyBoosting) 
 				stateMgr.state = KartState.DRIFTING;
-			else if(!value && stateMgr.state == KartState.DRIFTING)
+		} else { 
+			if(state == KartState.DRIFTING)
 				stateMgr.state = KartState.DRIVING;
 		}
 	}
 
-	public bool BoostInput {
-		get { return boosting; }
-		set {
-			if(value && !boosting && (boostAmount/maxBoost) >= requiredBoostPercentage) { 
-				boosting = true;
-				stateMgr.state = KartState.DRIVING;
-			} else if(!value) { 
-				if(boosting) boostAmount = 0;
-				boosting = false;
-			}
+	public void SetBoostInput(bool buttonDown) {
+		if(buttonDown && !boosting && (boostAmount/maxBoost) >= requiredBoostPercentage) { 
+			boosting = true;
+			stateMgr.state = KartState.DRIVING;
+		} else if(!buttonDown) { 
+			if(boosting) boostAmount = 0;
+			boosting = false;
 		}
 	}
 
 	/** The velocity magnitude tangential to the up vector */
-	public float TrackSpeed { 
-		get { 
-			if(rb == null) rb = GetComponent<Rigidbody>(); 
-			return new Vector3(rb.velocity.x, 0, rb.velocity.z).magnitude; 
-		} 
-	}
-
+	public float TrackSpeed { get { 
+		if(rb == null) rb = GetComponent<Rigidbody>(); 
+		return new Vector3(rb.velocity.x, 0, rb.velocity.z).magnitude; 
+	} }
 	public float CurrentMaxSpeed { get { return momentum == 1 ? (ActivelyBoosting ? maxBoostSpeed : maxSpeed) : maxSpeed/4f; } }
 	public float SpeedRatio { get { return TrackSpeed/CurrentMaxSpeed; } }
 
-	public bool SteeringWheelMatchesTurn { get { return Mathf.Sign(steeringWheelDirection) == Mathf.Sign(turn.x); } }
-	public bool SteeringWheelMatchesDrift { get { return Mathf.Sign(steeringWheelDirection) == Mathf.Sign(driftDirection); } }
+	public float GetSteeringWheelDirection() { return steeringWheelDirection; }
+	public bool SteeringWheelMatchesTurn { get { return Mathf.Sign(GetSteeringWheelDirection()) == Mathf.Sign(turn.x); } }
+	public bool SteeringWheelMatchesDrift { get { return Mathf.Sign(GetSteeringWheelDirection()) == Mathf.Sign(driftDirection); } }
+	/** Calculations related to forward/backward break at slow speeds, only do them above this speed. */
+	public bool CanChangeForward { get { return TrackSpeed > 1.25f; } }
 
-	public bool CanDriftEngage { get { return SpeedRatio >= driftEngageSpeedPercent && momentum == 1; } }
+	public bool CanDriftHop { get { return SpeedRatio >= driftHopSpeedPercent && momentum == 1; } }
+	public bool CanEngageDrift { get { return SpeedRatio >= driftEngageSpeedPercent; } }
 
+	public float GetBoostAmount() { return boostAmount; }
+	public bool IsBoosting() { return boosting; }
 	public bool ActivelyBoosting { get { return boosting && boostAmount > 0;} }
 	public float BoostRatio { get { return boostAmount/maxBoost; } }
-	
-	public float DriftTurnMultiplier { 
-		get { 
-			if(stateMgr.state == KartState.DRIFTING) {
-				float driftAgeRatio = stateMgr.timeInState/driftAge;
-				return SteeringWheelMatchesDrift ? 
-					Mathf.Lerp(turnMultiplierRangeDriftMatch.x, turnMultiplierRangeDriftMatch.y, 1-driftAgeRatio) :
-					Mathf.Lerp(turnMultiplierRangeDriftDiffer.x, turnMultiplierRangeDriftDiffer.y, 1-driftAgeRatio);
-			} else 
-				return 1f;
-		} 
-	}
 
-	private void DrawForces() { 
+	private void DrawRays() { 
 		// Debug.DrawLine(transform.position, transform.position + kartForward*3f, Color.blue, Time.deltaTime);
 	}
