--- Assets/Scripts/Kart/KartControllerV2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 50333dedf2a6ef3438011f98b7d5ef3c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
