+++ Assets/Scripts/Kart/KartStateManager.cs
@@ -47,7 +47,7 @@ public class KartStateManager : MonoBehaviour
 		/* Active state changes - Recieve requests */
 
 		/* Passive state changes */
-		if(state == KartState.DRIFTING && ((kc.Grounded() && kc.steeringWheelDirection == 0 && timeInState >= 0.15f) || !kc.CanDriftEngage)) { 
+		if(state == KartState.DRIFTING && kc.Grounded() && kc.GetSteeringWheelDirection() == 0 && timeInState >= 0.15f) { 
 			state = KartState.DRIVING;
 		}
 
@@ -57,5 +57,5 @@ public class KartStateManager : MonoBehaviour
 
 public enum KartState
 {
-	DRIVING, DRIFTING, REVERSING, COLLISION
+	DRIVING, DRIFTING, REVERSING
 }
